rename ValueKeyStatistic to StatisticKeyValue
