+++ src/components/index.ts
@@ -54,7 +54,6 @@ export { default as CopyableWrapper } from './CopyableWrapper.vue'
 export { default as CopyOverflowMenuItem } from './CopyOverflowMenuItem.vue'
 export { default as CronScheduleForm } from './CronScheduleForm.vue'
 export { default as CumulativeTaskRunsCard } from './CumulativeTaskRunsCard.vue'
-export { default as ValueKeyStatistic } from './ValueKeyStatistic.vue'
 export { default as DashboardWorkPoolsCard } from './DashboardWorkPoolsCard.vue'
 export { default as DateInput } from './DateInput.vue'
 export { default as DateRangeInput } from './DateRangeInput.vue'
@@ -223,6 +222,7 @@ export { default as StateIcon } from './StateIcon.vue'
 export { default as StateListItem } from './StateListItem.vue'
 export { default as StateNameSelect } from './StateNameSelect.vue'
 export { default as StateSelect } from './StateSelect.vue'
+export { default as StatisticKeyValue } from './StatisticKeyValue.vue'
 export { default as SubmitButton } from './SubmitButton.vue'
 export { default as TaskRunArtifacts } from './TaskRunArtifacts.vue'
 export { default as TaskRunDetails } from './TaskRunDetails.vue'
